Fix query parameters
Using the action_name, target_type and created_at columns in the WHERE
clause will limit the data returned by the api.  The created_at column,
which is mapped to the before and after gitlab api query parameters.
The code in gitlab/table_project_pipeline.go helped me set this up,
so thanks to Graza who wrote that.