Revert "Speed up the gitlab_issue table"
This reverts commit f682b6896334d29ee908c90bbde8788355aa9c9b.

Unfortunately the `GetIssueByID` call uses an api that requires `admin`
access: https://docs.gitlab.com/ee/api/issues.html#single-issue

The `GetIssue` call uses a non-privileged api, but it needs the
issue `iid`.  Since I was using this to speed up the `gitlab_user_event`
table which should have the issue `iid` in an "Isssue" event in the
`target_iid` column.  Unfortunately the underlying api doesn't seem
to return that.